RHODS: add full cluster cleanup

+++ testing/notebooks/prepare_sutest.sh
@@ -304,9 +304,8 @@ sutest_wait_rhods_launch() {
 
 sutest_cleanup() {
     switch_sutest_cluster
-    sutest_cleanup_ldap
 
-    skip_threshold=$(get_config tests.notebooks.cleanup.skip_if_le_than_users)
+    skip_threshold=$(get_config tests.notebooks.cleanup.on_exit.skip_if_le_than_users)
     user_count=$(get_config tests.notebooks.users.count)
     if [[ "$user_count" -le "$skip_threshold" ]]; then
         _info "Skip cluster cleanup (less that $skip_threshold users)"
@@ -333,6 +332,47 @@ sutest_cleanup() {
     else
         ./run_toolbox.py from_config cluster set_scale --prefix "sutest" --suffix "cleanup" > /dev/null
     fi
+
+    if test_config tests.notebooks.cleanup.on_exit.sutest.uninstall_rhods; then
+        echo "$(date +%H:%M:%S) Uninstall RHODS ..."
+        oc create configmap delete-self-managed-odh -n redhat-ods-operator
+        oc label configmap/delete-self-managed-odh api.openshift.com/addon-managed-odh-delete=true -n redhat-ods-operator
+        for project in redhat-ods-applications redhat-ods-monitoring rhods-notebooks; do
+            local current_remaining_time=$((5 * 60)) # 5 minutes
+            while oc get ns "$project" &> /dev/null; do
+                sleep 1
+                current_remaining_time=$((current_remaining_time - 1))
+                if [[ "$current_remaining_time" == 0 ]]; then
+                    _error "$(date +%H:%M:%S) RHODS deletion failed :/"
+                    exit 1 # shouldn't be reached
+                fi
+
+            done
+            echo "$(date +%H:%M:%S) Namespace $project no longer exists"
+        done
+        oc delete ns redhat-ods-operator
+        echo "$(date +%H:%M:%S) Uninstall of RHODS done."
+    fi
+
+    if test_config tests.notebooks.cleanup.on_exit.sutest.delete_test_namespaces; then
+        echo "Deleting the notebook performance namespace"
+       local notebook_performance_namespace=$(get_command_arg namespace rhods benchmark_notebook_performance)
+       oc delete namespace --ignore-not-found \
+          "$notebook_performance_namespace"
+    fi
+
+    if test_config tests.notebooks.cleanup.on_exit.sutest.uninstall_ldap; then
+        echo "Uninstaling LDAP IDP"
+        sutest_cleanup_ldap
+    fi
+
+    if test_config tests.notebooks.cleanup.on_exit.sutest.remove_dsg_notebook_dedicated_toleration; then
+        dedicated="{value: ''}" # delete the toleration/node-selector annotations, if it exists
+        echo "Removing the DSG notebook toleration"
+
+        ./run_toolbox.py from_config cluster set_project_annotation --prefix sutest --suffix node_selector --extra "$dedicated" > /dev/null
+        ./run_toolbox.py from_config cluster set_project_annotation --prefix sutest --suffix toleration --extra "$dedicated" > /dev/null
+    fi
 }
 
 sutest_cleanup_ldap() {
@@ -358,6 +398,11 @@ sutest_cleanup_rhods() {
     local user_prefix=$(get_config ldap.users.prefix)
 
     switch_sutest_cluster
+
+    if ! oc get ns redhat-ods-applications > /dev/null; then
+        _info "RHODS not installed, cannot clean it up."
+        return
+    fi
 
     oc delete namespaces -lopendatahub.io/dashboard=true >/dev/null
     # delete any leftover namespace (if the label ^^^ wasn't properly applied)
